refactor: getting validator singleton instance
1. Use sync.Once for getting singleton instance
2. Initialize validator in main function instead of init
3. Add more comments for understanding code
